Avalon SDK for direct mode Worker related API
