+++ src/common/interfaces/work_order_receipt.py
