+++ src/common/interfaces/worker_registry.py
