+++ src/common/interfaces/worker_registry_list.py
