+++ src/internal/handler/direct/__init__.py
@@ -12,4 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-__all__ = []
+
+__all__ = [ ]
+
